refactor: 로또 숫자 클래스명 변경, 유효하지 않은 숫자 예외 추가
